requirements.txt
# Python dependencies for reproducing analyses in:
# "Human Coordination Shapes the Promise and Limits of Autonomous Agents"

# Analysis pipeline (run_data_analysis.py, stages 06-11)
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
cycler>=0.11

# Collection pipeline (run_data_processing.py, stages 01-05)
requests>=2.28
tqdm>=4.64

# Bot categorization pipeline (scripts/bot_*.py)
# Only needed to re-run the LLM annotation; the final labels are already in data/.
openai>=1.0
anthropic>=0.20
instructor>=1.0        # structured outputs for Anthropic
google-generativeai>=0.8
tenacity>=8.0
scikit-learn>=1.3      # bot_label_evaluation.py (accuracy metrics)
